added regression tests

+++ test.py
@@ -8,13 +8,24 @@ class TestHarness():
     def __init__(self):
         """Initialize the test harness"""
         self.orig_stdout = sys.stdout
-        self.filename = "test_results.txt"
-        os.remove(self.filename)
+        
+        self.filename = "results.txt"
+        if(os.path.isfile(self.filename)):
+            os.remove(self.filename)
         self.f = open(self.filename,"w")
         sys.stdout = self.f
         self.owd   = os.getcwd()
-        self.file_locations = [r".\src\python",r".\src\fortran"]
+        
+        #Identify module locations
+        self.module_location = r".\src\python"
         self.fformat = "./*.py"
+        
+        #Add the module locations to the system path for now
+        sys.path.append(os.path.abspath(self.module_location))
+        
+        #Identify regression tests
+        self.regression_tests = [r"tests/regression_tests/const_u/const_u.inp",\
+                                 r"tests/regression_tests/linear_u/linear_u.inp"]
         
     def __del__(self):
         print "======================================\n"+\
@@ -40,16 +51,15 @@ class TestHarness():
               "| functions.                         |\n"+\
               "======================================\n"
         
-        os.chdir(os.path.join(self.owd,self.file_locations[0]))
+        #Run the unit tests for modules located in the module location
+        os.chdir(os.path.join(self.owd,self.module_location))
         files = glob.glob(self.fformat)
         
-        #Parse through files
+        #Run unit tests on modules
         for f in files:
             print "\nTesting: {0}\n".format(f)
             os.system("python {0} -v".format(f))
-            
-        os.chdir(self.owd)
-        
+
         print "======================================\n"+\
               "|                                    |\n"+\
               "|        Run regression tests        |\n"+\
@@ -60,14 +70,16 @@ class TestHarness():
               "| against the code as a whole.       |\n"+\
               "======================================\n"
         
-        os.chdir(os.path.join(self.owd,self.file_locations[1]))
-        files = glob.glob(self.fformat)
-        
-        for f in files:
-            print "\nTesting: {0}\n".format(f)
-            os.system("python {0} -v".format(f))
-            
-        os.chdir(self.owd)
+        import fea_driver
+        for f in self.regression_tests:
+            try:
+                sys.stdout = self.orig_stdout                           #Change the output to the screen instead of the results file
+                fea_driver.run_finite_element_model(os.path.abspath(f)) #Run the finite element model indicated
+            except:
+                print "!!! Error in execution !!!\nError: {0}".format(sys.exc_info()[0]) #Raise exception
+                sys.stdout = self.f                                                      #Set the standard out back to the results file
+                print "!!! Error in execution !!!\nError: {0}".format(sys.exc_info()[0]) #Repeat the error message into the file
+        os.chdir(self.owd) #Reset the working directory
